# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project API",
        "description": "Project API",
        "version": null
    },
    "servers": [
        {
            "url": "http://project.local/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Gets list of all users",
                "description": "This API Endpoint returns lists of all existing users.",
                "operationId": "UsersList",
                "responses": {
                    "200": {
                        "description": "Successfull Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Creates a new user",
                "description": "This API Endpoint creates a new user.",
                "operationId": "UsersCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfull Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "ResponseUser",
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/users/{userId}/groups": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get groups of a user",
                "description": "This API Endpoint get groups of a user.",
                "operationId": "UsersGroupsGet",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Group"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/{userId}/groups/{groupId}/revoke": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Removes a user from a group",
                "description": "This API Endpoint removes users from groups.",
                "operationId": "UsersGroupsRevoke",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Revoked"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Group": {
                "description": "Group Resource",
                "properties": {
                    "id": {
                        "description": "Group ID",
                        "type": "integer",
                        "example": 0
                    },
                    "name": {
                        "description": "Group name",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Developers"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "description": "User Resource",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 0
                    },
                    "roleId": {
                        "description": "Role ID",
                        "type": "integer",
                        "example": 0
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "maxLength": 64,
                        "example": "John"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "token": []
        }
    ]
}